Implemented HP/MP limit break using GameDoubleHealth (add 28). Controlled via config.lua flag.

+++ src/creature.cpp
@@ -662,7 +662,7 @@ void Creature::onDeath()
 
 	const int64_t timeNow = OTSYS_TIME();
 	const int64_t inFightTicks = getInteger(ConfigManager::PZ_LOCKED);
-	int32_t mostDamage = 0;
+	int64_t mostDamage = 0;
 	std::map<Creature*, uint64_t> experienceMap;
 	for (const auto& it : damageMap) {
 		if (Creature* attacker = g_game.getCreatureByID(it.first)) {
@@ -792,14 +792,14 @@ bool Creature::hasBeenAttacked(uint32_t attackerId)
 
 Item* Creature::getCorpse(Creature*, Creature*) { return Item::CreateItem(getLookCorpse()); }
 
-void Creature::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*/)
+void Creature::changeHealth(int64_t healthChange, bool sendHealthChange /* = true*/)
 {
-	int32_t oldHealth = health;
+	int64_t oldHealth = health;
 
 	if (healthChange > 0) {
-		health += std::min<int32_t>(healthChange, getMaxHealth() - health);
+		health += std::min<int64_t>(healthChange, getMaxHealth() - health);
 	} else {
-		health = std::max<int32_t>(0, health + healthChange);
+		health = std::max<int64_t>(0, health + healthChange);
 	}
 
 	if (sendHealthChange && oldHealth != health) {
@@ -811,7 +811,7 @@ void Creature::changeHealth(int32_t healthChange, bool sendHealthChange /* = tru
 	}
 }
 
-void Creature::gainHealth(Creature* healer, int32_t healthGain)
+void Creature::gainHealth(Creature* healer, int64_t healthGain)
 {
 	changeHealth(healthGain);
 	if (healer) {
@@ -819,7 +819,7 @@ void Creature::gainHealth(Creature* healer, int32_t healthGain)
 	}
 }
 
-void Creature::drainHealth(Creature* attacker, int32_t damage)
+void Creature::drainHealth(Creature* attacker, int64_t damage)
 {
 	changeHealth(-damage, false);
 
@@ -830,7 +830,7 @@ void Creature::drainHealth(Creature* attacker, int32_t damage)
 	}
 }
 
-BlockType_t Creature::blockHit(Creature* attacker, CombatType_t combatType, int32_t& damage,
+BlockType_t Creature::blockHit(Creature* attacker, CombatType_t combatType, int64_t& damage,
                                bool checkDefense /* = false */, bool checkArmor /* = false */, bool /* field = false */,
                                bool /* ignoreResistances = false */)
 {
@@ -1042,8 +1042,8 @@ bool Creature::setFollowCreature(Creature* creature)
 
 double Creature::getDamageRatio(Creature* attacker) const
 {
-	uint32_t totalDamage = 0;
-	uint32_t attackerDamage = 0;
+	uint64_t totalDamage = 0;
+	uint64_t attackerDamage = 0;
 
 	for (const auto& it : damageMap) {
 		const CountBlock_t& cb = it.second;
@@ -1065,7 +1065,7 @@ uint64_t Creature::getGainedExperience(Creature* attacker) const
 	return std::floor(getDamageRatio(attacker) * getLostExperience());
 }
 
-void Creature::addDamagePoints(Creature* attacker, int32_t damagePoints)
+void Creature::addDamagePoints(Creature* attacker, int64_t damagePoints)
 {
 	if (damagePoints <= 0) {
 		return;
@@ -1143,7 +1143,7 @@ void Creature::onAttacked()
 	//
 }
 
-void Creature::onAttackedCreatureDrainHealth(Creature* target, int32_t points)
+void Creature::onAttackedCreatureDrainHealth(Creature* target, int64_t points)
 {
 	target->addDamagePoints(this, points);
 }
