+++ src/tools.h
@@ -31,8 +31,8 @@ IntegerVector vectorAtoi(const std::vector<std::string_view>& stringVector);
 constexpr bool hasBitSet(uint32_t flag, uint32_t flags) { return (flags & flag) != 0; }
 
 std::mt19937& getRandomGenerator();
-int32_t uniform_random(int32_t minNumber, int32_t maxNumber);
-int32_t normal_random(int32_t minNumber, int32_t maxNumber);
+int64_t uniform_random(int64_t minNumber, int64_t maxNumber);
+int64_t normal_random(int64_t minNumber, int64_t maxNumber);
 bool boolean_random(double probability = 0.5);
 
 Position getNextPosition(Direction direction, Position pos);
